Fix ? Maintenant, oui

+++ 3d-printer-error-detector/tools.py
@@ -30,13 +30,16 @@ def capture_image():
 
             # Remove old images
             files = os.listdir(static_folder)
-            for file in files:
-                file_path = os.path.join(static_folder, file)
-                if os.path.isfile(file_path):
-                    os.remove(file_path)
+            files.sort()
+            number_of_files = len(files)
+            if number_of_files > settings.NUMBER_OF_IMAGES_RETAINED - 1:
+                for file in files[:number_of_files - settings.NUMBER_OF_IMAGES_RETAINED + 1]:
+                    file_path = os.path.join(static_folder, file)
+                    if os.path.isfile(file_path):
+                        os.remove(file_path)
 
-            image_name = str(int(time.time())) + '.jpg'
-            image_path = os.path.join(static_folder, image_name)
+            settings.image_name = str(int(time.time())) + '.jpg'
+            image_path = os.path.join(static_folder, settings.image_name)
             cv2.imwrite(image_path, frame)
             settings.image_path = image_path
             cap.release()
